having the stuff in a state

+++ rename_files.sh
@@ -9,17 +9,21 @@ instaloader --filename-pattern={date_utc} profile sidhochwind
 cd sidhochwind 
 
 # delete double image
-rm "2021-02-06_16-59-45_2.jpg"
-# replace the darling
-mv "../killed-darlings/2014-05-06_15-13-35.jpg" "2013-12-25_12-54-41.jpg"
+mv "2021-02-06_16-59-45_2.jpg" "../killed-darlings/2021-02-06_16-59-45_2.jpg"
+mv "2017-10-20_19-28-35_1.jpg" "../killed-darlings/2017-10-20_19-28-35_1.jpg"
+
+# replace the ugly one with an old darling
+cp "../killed-darlings/2014-05-06_15-13-35.jpg" "2013-12-25_12-54-41.jpg"
 
 # no idea why we have sometimes too much or to less images but well no time now...
 # 3rd year -2
 mv "2015-10-08_12-26-30.jpg" "../killed-darlings/2015-10-08_12-26-30.jpg"
 mv "2015-10-12_07-28-11.jpg" "../killed-darlings/2015-10-12_07-28-11.jpg"
-# 4rd year +1
-mv "../killed-darlings/2018-07-04_08-58-41.jpg" "2018-07-04_08-58-41.jpg"
-# 5th year fine
+# 4rd year +2
+cp "../killed-darlings/2017-05-17_15-17-27.jpg" "2017-05-17_15-17-27.jpg"
+cp "../killed-darlings/2017-05-28_12-52-26.jpg" "2017-05-28_12-52-26.jpg"
+# 5th year -1
+mv "2018-07-04_08-58-41.jpg" "../killed-darlings/2018-07-04_08-58-41.jpg"
 # 6th year -1
 mv "2018-12-28_10-55-00_2.jpg" "../killed-darlings/2018-12-28_10-55-00_2.jpg"
 # 7th year -1
@@ -52,7 +56,11 @@ do
 
         # rename the files
         cp "$ori_jpg_name" "$jpg_name"
-        cp "$ori_txt_name" "$txt_name"
+        if [ -e "$ori_txt_name" ]; then
+            cp "$ori_txt_name" "$txt_name"
+        else
+             touch "$txt_name"
+        fi
 
         if [ $counter = 359 ]; then
             counter=0
